Carve static prompt-token estimation and pure budget math out of compaction.py
- history/prompt_tokens.py now owns the ~370 lines of static-token
  estimation (estimators, Agno tool preparation, tool-schema payload
  logging, compute_prompt_token_breakdown): none of it touched scopes,
  sessions, or summaries, and it dragged agno.team internals and
  tool_schema_cache imports into the compaction module.
- The three pure budget helpers (effective threshold, budget clamp,
  compaction runtime settings) move into policy.py, which previously
  imported the 1,500-line compaction module just for them; the two with
  no external callers become private.
- tach: new prompt_tokens module entry; policy no longer depends on
  compaction; tool_schema_cache visibility follows the moved code.
