+++ src/mindroom/history/__init__.py
@@ -1,20 +1,20 @@
 """Persisted history compaction helpers."""
 
 from mindroom.history import agno_team_patch
-from mindroom.history.compaction import (
-    StaticTokenEstimator,
-    agent_static_token_estimator,
-    agent_tool_definition_payloads_for_logging,
-    compute_prompt_token_breakdown,
-    normalize_compaction_budget_tokens,
-    team_static_token_estimator,
-    team_tool_definition_payloads_for_logging,
-)
 from mindroom.history.manual import request_compaction_before_next_reply
 from mindroom.history.policy import (
     context_budget_after_reserve,
     manual_compaction_unavailable_message,
+    normalize_compaction_budget_tokens,
     resolve_history_execution_plan,
+)
+from mindroom.history.prompt_tokens import (
+    StaticTokenEstimator,
+    agent_static_token_estimator,
+    agent_tool_definition_payloads_for_logging,
+    compute_prompt_token_breakdown,
+    team_static_token_estimator,
+    team_tool_definition_payloads_for_logging,
 )
 from mindroom.history.runtime import (
     HistoryPreparationInputs,
